Remove syslog-ng-module-extra from module list
Re-introducing syslog-ng-module-extra had an un-wanted side effect that
although is just a transitional package now, it is part of the suggested
package list for syslog-ng-core and part of the recommended package list in
case of syslog-ng.

As this wasn't intentional and mopst probably this dependency is what brakes
the autopkgtest (although it worked for me), it is better to remove it.